Refacto 1st round : modularize versioning
