+++ plugin.py
@@ -42,7 +42,11 @@ from qgis.core import QgsCredentials, QgsDataSourceURI, QgsMapLayerRegistry, \
     QgsRuleBasedRendererV2
 from PyQt4.QtCore import *
 
-from . import versioning
+from .versioningDB.utils import Db, mem_field_names_types, get_actual_pk, get_pg_users_list
+from .versioningDB.sp_versioning import spVersioning
+from .versioningDB.pg_versioning import pgVersioning
+from .versioningDB import versioning
+
 
 # Deactivate stdout (like output of print statements) because windows
 # causes occasional "IOError [Errno 9] File descriptor error"
@@ -101,6 +105,9 @@ class Plugin(QObject):
             self.legend = self.iface.mainWindow().findChild( QTreeView, 'theLayerTreeView')
             self.legend.clicked.connect(self.on_legend_click)
 
+        self.spversioning = spVersioning()
+        self.pgversioning = pgVersioning()
+        
     def enable_diffmode(self):
         '''This function enables the diffmode checkbox iif the number of checked
         revision == 2.  The reason is that we want to apply diff styles to
@@ -164,22 +171,22 @@ class Plugin(QObject):
         srid = str(QgsDataSourceURI(pg_layer.source()).srid())
         if pg_layer.wkbType() == QGis.WKBPoint:
             #print "Layer \"" + pg_layer.name()+ "\" is a point layer"
-            mem_uri = "Point?crs=epsg:" + srid +"&" + versioning.mem_field_names_types(pg_layer) + "&index=yes"
+            mem_uri = "Point?crs=epsg:" + srid +"&" + mem_field_names_types(pg_layer) + "&index=yes"
         if pg_layer.wkbType()==QGis.WKBLineString:
             #print "Layer \"" + pg_layer.name()+ "\" is a linestring layer"
-            mem_uri = "LineString?crs=epsg:" + srid +"&" + versioning.mem_field_names_types(pg_layer) + "&index=yes"
+            mem_uri = "LineString?crs=epsg:" + srid +"&" + mem_field_names_types(pg_layer) + "&index=yes"
         if pg_layer.wkbType() == QGis.WKBPolygon:
             #print "Layer \"" + pg_layer.name()+ "\" is a polygon layer"
-            mem_uri = "Polygon?crs=epsg:" + srid +"&" + versioning.mem_field_names_types(pg_layer) + "&index=yes"
+            mem_uri = "Polygon?crs=epsg:" + srid +"&" + mem_field_names_types(pg_layer) + "&index=yes"
         if pg_layer.wkbType() == QGis.WKBMultiPoint:
             #print "Layer \"" + pg_layer.name()+ "\" is a multi-point layer"
-            mem_uri = "MultiPoint?crs=epsg:" + srid +"&" + versioning.mem_field_names_types(pg_layer) + "&index=yes"
+            mem_uri = "MultiPoint?crs=epsg:" + srid +"&" + mem_field_names_types(pg_layer) + "&index=yes"
         if pg_layer.wkbType()==QGis.WKBMultiLineString:
             #print "Layer \"" + pg_layer.name()+ "\" is a multi-linestring layer"
-            mem_uri = "MultiLineString?crs=epsg:" + srid +"&" + versioning.mem_field_names_types(pg_layer) + "&index=yes"
+            mem_uri = "MultiLineString?crs=epsg:" + srid +"&" + mem_field_names_types(pg_layer) + "&index=yes"
         if pg_layer.wkbType()==QGis.WKBMultiPolygon:
             #print "Layer \"" + pg_layer.name()+ "\" is a multi-polygon layer"
-            mem_uri = "MultiPolygon?crs=epsg:" + srid +"&" + versioning.mem_field_names_types(pg_layer) + "&index=yes"
+            mem_uri = "MultiPolygon?crs=epsg:" + srid +"&" + mem_field_names_types(pg_layer) + "&index=yes"
         return mem_uri
 
     def pg_conn_info(self):
@@ -221,7 +228,7 @@ class Plugin(QObject):
 
     def get_pg_users_list(self):
         #get list of pg users to populate combobox used in commit msg dialog
-        pg_users_list = versioning.get_pg_users_list(self.pg_conn_info())
+        pg_users_list = get_pg_users_list(self.pg_conn_info())
         #print "pg_users_list = " + str(pg_users_list)
         self.q_commit_msg_dlg.pg_users_combobox.addItems (pg_users_list)
 
@@ -304,7 +311,7 @@ class Plugin(QObject):
         if layer.providerType() == "spatialite":
             rev = 0
             try:
-                rev = versioning.revision( uri.database() )
+                rev = self.spversioning.revision( uri.database() )
             except:
                 self.current_layers = []
                 self.info.setText("Versioning : the selected group is not a working copy")
@@ -326,8 +333,8 @@ class Plugin(QObject):
                 # check if it's a working copy
                 rev = 0
                 try:
-                    rev = versioning.pg_revision( self.pg_conn_info(),
-                                                       uri.schema() )
+                    rev = self.pgversioning.revision( [self.pg_conn_info(),
+                                                       uri.schema()] )
                     selection_type = 'working copy'
                     self.info.setText( uri.database()+' '+uri.schema()
                             +' <b>working rev</b>='+str(rev) )
@@ -468,7 +475,7 @@ class Plugin(QObject):
             print 'aborted'
             return
 
-        pcur = versioning.Db( psycopg2.connect(self.pg_conn_info()) )
+        pcur = Db( psycopg2.connect(self.pg_conn_info()) )
         pcur.execute("SELECT * FROM "+schema+".revisions "
             "WHERE branch = '"+branch+"'")
         if pcur.fetchone():
@@ -532,7 +539,7 @@ class Plugin(QObject):
         self.q_view_dlg.diffmode_chk.setCheckState(Qt.Unchecked)
         self.q_view_dlg.diffmode_chk.setEnabled(False)
 
-        pcur = versioning.Db( psycopg2.connect(self.pg_conn_info()) )
+        pcur = Db( psycopg2.connect(self.pg_conn_info()) )
         pcur.execute("SELECT rev, author, date::timestamp(0), branch, commit_msg "
             "FROM "+schema+".revisions ORDER BY rev ASC")
         revs = pcur.fetchall()
@@ -735,7 +742,7 @@ class Plugin(QObject):
         uri = QgsDataSourceURI(layer.source())
 
         if layer.providerType() == "spatialite":
-            unresolved = versioning.unresolved_conflicts( uri.database() )
+            unresolved = self.spversioning.unresolved_conflicts( [uri.database()] )
             for cflt in unresolved:
                 table = cflt+"_conflicts"
                 if not QgsMapLayerRegistry.instance().mapLayersByName(table):
@@ -746,8 +753,8 @@ class Plugin(QObject):
                             " key=\"OGC_FID\" table=\""+table+"\" "+
                             geom,table,'spatialite')
         else: #postgres
-            unresolved = versioning.pg_unresolved_conflicts(
-                    uri.connectionInfo(), uri.schema() )
+            unresolved = self.pgversioning.unresolved_conflicts(
+                    [uri.connectionInfo(), uri.schema()] )
             for cflt in unresolved:
                 table = cflt+"_conflicts"
                 if not QgsMapLayerRegistry.instance().mapLayersByName(table):
@@ -787,19 +794,19 @@ class Plugin(QObject):
         late_by = 0
 
         if layer.providerType() == "spatialite":
-            late_by = versioning.late(
-                    uri.database(), self.pg_conn_info() )
+            late_by = self.spversioning.late(
+                    [uri.database(), self.pg_conn_info()] )
         else: # postgres
-            late_by = versioning.pg_late(
-                    self.pg_conn_info(), uri.schema() )
+            late_by = self.pgversioning.late(
+                    [self.pg_conn_info(), uri.schema()] )
         if late_by:
             if layer.providerType() == "spatialite":
-                versioning.update( uri.database(), self.pg_conn_info() )
-                rev = versioning.revision( uri.database() )
+                self.spversioning.update( [uri.database(), self.pg_conn_info()] )
+                rev = self.spversioning.revision( uri.database() )
             else: # postgres
-                versioning.pg_update( uri.connectionInfo(), uri.schema() )
-                rev = versioning.pg_revision(
-                        uri.connectionInfo(), uri.schema() )
+                self.pgversioning.update( [uri.connectionInfo(), uri.schema()] )
+                rev = self.pgversioning.revision(
+                        [uri.connectionInfo(), uri.schema()] )
 
             # Force refresh of map
             if self.iface.mapCanvas().isCachingEnabled():
@@ -820,10 +827,10 @@ class Plugin(QObject):
                 "Now up to date with remote revision "+str(rev-1)+".")
         else:
             if layer.providerType() == "spatialite":
-                rev = versioning.revision( uri.database() )
+                rev = self.spversioning.revision( uri.database() )
             else: # postgres
-                rev = versioning.pg_revision(
-                        uri.connectionInfo(), uri.schema() )
+                rev = self.pgversioning.revision(
+                        [uri.connectionInfo(), uri.schema()] )
             QMessageBox.information( self.iface.mainWindow(), "Info","Working "
             "copy already up to date with remote revision "+str(rev-1)+".")
 
@@ -885,7 +892,7 @@ class Plugin(QObject):
             uri = QgsDataSourceURI(layer.source())
 
             # Get actual PK fror corresponding table
-            actual_table_pk = versioning.get_actual_pk( uri,self.pg_conn_info() )
+            actual_table_pk = get_actual_pk( uri,self.pg_conn_info() )
             #print "Actual table pk = " + actual_table_pk
 
             layer_selected_features_ids = [f[actual_table_pk] for f in layer.selectedFeatures()]
@@ -927,7 +934,7 @@ class Plugin(QObject):
             os.remove(filename)
 
         print "checking out ", tables_for_conninfo, " from ",uri.connectionInfo()
-        versioning.checkout( self.pg_conn_info(),
+        self.spversioning.checkout( self.pg_conn_info(),
                 tables_for_conninfo, filename, user_selected_features )
 
         # add layers from offline version
@@ -961,7 +968,7 @@ class Plugin(QObject):
             uri = QgsDataSourceURI(layer.source())
 
             # Get actual PK fror corresponding table
-            actual_table_pk = versioning.get_actual_pk( uri,self.pg_conn_info() )
+            actual_table_pk = get_actual_pk( uri,self.pg_conn_info() )
             #print "Actual table pk = " + actual_table_pk
 
             layer_selected_features_ids = [f[actual_table_pk] for f in layer.selectedFeatures()]
@@ -1031,7 +1038,7 @@ class Plugin(QObject):
             "<b>lowercase</b> letters, underscore or digits.", duration=10)
             return
         print "checking out ", tables_for_conninfo, " from ", uri.connectionInfo()
-        versioning.pg_checkout( self.pg_conn_info(),
+        self.pgversioning.checkout( self.pg_conn_info(),
                 tables_for_conninfo, working_copy_schema, user_selected_features )
 
         # add layers from offline version
@@ -1062,11 +1069,11 @@ class Plugin(QObject):
 
         late_by = 0
         if layer.providerType() == "spatialite":
-            late_by = versioning.late(
-                    uri.database(), self.pg_conn_info() )
+            late_by = self.spversioning.late(
+                    [uri.database(), self.pg_conn_info()] )
         else:#postgres
-            late_by = versioning.pg_late(
-                    self.pg_conn_info(), uri.schema() )
+            late_by = self.pgversioning.late(
+                    [self.pg_conn_info(), uri.schema()] )
 
         if late_by:
             QMessageBox.warning(self.iface.mainWindow(), "Warning",
@@ -1110,14 +1117,14 @@ class Plugin(QObject):
         nb_of_updated_layer = 0
         rev = 0
         if layer.providerType() == "spatialite":
-            nb_of_updated_layer = versioning.commit( uri.database(),
-                    commit_msg, self.pg_conn_info(),commit_pg_user )
-            rev = versioning.revision(uri.database())
+            nb_of_updated_layer = self.spversioning.commit( [uri.database(), self.pg_conn_info()], 
+                    commit_msg, commit_pg_user )
+            rev = self.spversioning.revision(uri.database())
         else: # postgres
-            nb_of_updated_layer = versioning.pg_commit(
-                    uri.connectionInfo(), uri.schema(), commit_msg )
-            rev = versioning.pg_revision(
-                    uri.connectionInfo(), uri.schema())
+            nb_of_updated_layer = self.pgversioning.commit(
+                    [uri.connectionInfo(), uri.schema()], commit_msg )
+            rev = self.pgversioning.revision(
+                    [uri.connectionInfo(), uri.schema()])
 
         if nb_of_updated_layer:
             #self.iface.messageBar().pushMessage("Info",
