+++ test/bug_in_branch_after_commit.py
@@ -1,35 +1,45 @@
-#!/usr/bin/python
-from .. import versioning
+#!/usr/bin/env python2
+from __future__ import absolute_import
+import sys
+sys.path.insert(0, '..')
+
+from versioningDB import versioning
 from pyspatialite import dbapi2
-import psycopg2
+from versioningDB.sp_versioning import spVersioning
+from versioningDB.utils import Db
 import os
-import shutil
 import tempfile
 
+PGUSER = 'postgres'
+HOST = '127.0.0.1'
+
+pg_conn_info = "dbname=epanet_test_db host="+HOST+" user="+PGUSER
+
 if __name__ == "__main__":
+    spversioning = spVersioning()
     test_data_dir = os.path.dirname(os.path.realpath(__file__))
     tmp_dir = tempfile.gettempdir()
 
+
     # create the test database
+    os.system("dropdb --if-exists -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("createdb -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -c 'CREATE EXTENSION postgis'")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -f "+test_data_dir+"/epanet_test_db.sql")
 
-    os.system("dropdb epanet_test_db")
-    os.system("createdb epanet_test_db")
-    os.system("psql epanet_test_db -c 'CREATE EXTENSION postgis'")
-    os.system("psql epanet_test_db -f "+test_data_dir+"/epanet_test_db_unversioned.sql")
-
-    versioning.historize("dbname=epanet_test_db","epanet")
+    versioning.historize(pg_conn_info,"epanet")
 
     # try the update
     wc = tmp_dir+"/bug_in_branch_after_commit_wc.sqlite"
     if os.path.isfile(wc): os.remove(wc) 
-    versioning.checkout("dbname=epanet_test_db", ['epanet_trunk_rev_head.junctions', 'epanet_trunk_rev_head.pipes'], wc)
+    spversioning.checkout(pg_conn_info, ['epanet_trunk_rev_head.junctions', 'epanet_trunk_rev_head.pipes'], wc)
 
-    scur = versioning.Db( dbapi2.connect( wc ) )
+    scur = Db( dbapi2.connect( wc ) )
 
     scur.execute("SELECT * FROM pipes")
     scur.execute("UPDATE pipes_view SET length = 1 WHERE OGC_FID = 1")
     scur.commit()
 
-    versioning.commit(wc,'test', "dbname=epanet_test_db" )
+    spversioning.commit([wc, pg_conn_info],'test' )
 
-    versioning.add_branch("dbname=epanet_test_db","epanet","mybranch","add 'branch")
+    versioning.add_branch(pg_conn_info,"epanet","mybranch","add 'branch")
