+++ test/issue287_test.py
@@ -1,25 +1,34 @@
-from .. import versioning
-from pyspatialite import dbapi2
-import psycopg2
+#!/usr/bin/env python2
+from __future__ import absolute_import
+import sys
+sys.path.insert(0, '..')
+
+from versioningDB.sp_versioning import spVersioning
 import os
 import shutil
 import tempfile
 
+PGUSER = 'postgres'
+HOST = '127.0.0.1'
+
+pg_conn_info = "dbname=epanet_test_db host="+HOST+" user="+PGUSER
+
 def test():
+    spversioning = spVersioning()
+    
     test_data_dir = os.path.dirname(os.path.realpath(__file__))
     tmp_dir = tempfile.gettempdir()
 
     # create the test database
-
-    os.system("dropdb epanet_test_db")
-    os.system("createdb epanet_test_db")
-    os.system("psql epanet_test_db -c 'CREATE EXTENSION postgis'")
-    os.system("psql epanet_test_db -f "+test_data_dir+"/issue287_pg_dump.sql")
+    os.system("dropdb --if-exists -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("createdb -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -c 'CREATE EXTENSION postgis'")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -f "+test_data_dir+"/epanet_test_db.sql")
 
     # try the update
     shutil.copyfile(test_data_dir+"/issue287_wc.sqlite", tmp_dir+"/issue287_wc.sqlite")
-    versioning.update(tmp_dir+"/issue287_wc.sqlite", "dbname=epanet_test_db")
-    versioning.commit(tmp_dir+"/issue287_wc.sqlite", "test message", "dbname=epanet_test_db")
+    spversioning.update([ tmp_dir+"/issue287_wc.sqlite", pg_conn_info] )
+    spversioning.commit([tmp_dir+"/issue287_wc.sqlite", pg_conn_info], "test message")
 
 if __name__ == "__main__":
     test()
