+++ test/issue357_test.py
@@ -1,38 +1,48 @@
-#!/usr/bin/python
-from .. import versioning
+#!/usr/bin/env python2
+from __future__ import absolute_import
+import sys
+sys.path.insert(0, '..')
+
 from pyspatialite import dbapi2
+from versioningDB.sp_versioning import spVersioning
+from versioningDB.utils import Db
 import psycopg2
 import os
 import shutil
 import tempfile
 
+PGUSER = 'postgres'
+HOST = '127.0.0.1'
+
+pg_conn_info = "dbname=epanet_test_db host="+HOST+" user="+PGUSER
+
 def test():
+    spversioning = spVersioning()
     test_data_dir = os.path.dirname(os.path.realpath(__file__))
     tmp_dir = tempfile.gettempdir()
 
     # create the test database
-
-    os.system("dropdb epanet_test_db")
-    os.system("createdb epanet_test_db")
-    os.system("psql epanet_test_db -c 'CREATE EXTENSION postgis'")
-    os.system("psql epanet_test_db -f "+test_data_dir+"/epanet_test_db.sql")
+    os.system("dropdb --if-exists -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("createdb -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -c 'CREATE EXTENSION postgis'")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -f "+test_data_dir+"/epanet_test_db.sql")
 
     # try the update
     wc = [tmp_dir+"/issue357_wc0.sqlite", tmp_dir+"/issue357_wc1.sqlite"]
     for f in wc:
         if os.path.isfile(f): os.remove(f) 
-        versioning.checkout("dbname=epanet_test_db", ['epanet_trunk_rev_head.junctions', 'epanet_trunk_rev_head.pipes'], f)
+        spversioning.checkout(pg_conn_info, ['epanet_trunk_rev_head.junctions', 'epanet_trunk_rev_head.pipes'], f)
 
     scur = []
-    for f in wc: scur.append(versioning.Db( dbapi2.connect( f ) ))
+    for f in wc: scur.append(Db( dbapi2.connect( f ) ))
 
     scur[0].execute("INSERT INTO pipes_view(id, start_node, end_node, GEOMETRY) VALUES ('2','1','2',GeomFromText('LINESTRING(1 1,0 1)',2154))")
     scur[0].execute("INSERT INTO pipes_view(id, start_node, end_node, GEOMETRY) VALUES ('3','1','2',GeomFromText('LINESTRING(1 -1,0 1)',2154))")
     scur[0].commit()
 
 
-    versioning.commit( wc[0], 'commit 1 wc0', "dbname=epanet_test_db")
-    versioning.update( wc[1], "dbname=epanet_test_db" )
+    spversioning.commit( [wc[0], pg_conn_info], 'commit 1 wc0')
+    spversioning.update( [wc[1], pg_conn_info] )
 
     scur[0].execute("UPDATE pipes_view SET length = 1")
     scur[0].commit()
@@ -40,7 +50,7 @@ def test():
     scur[1].execute("UPDATE pipes_view SET length = 3")
     scur[1].commit()
 
-    versioning.commit( wc[0], "commit 2 wc0", "dbname=epanet_test_db" )
+    spversioning.commit( [wc[0], pg_conn_info], "commit 2 wc0" )
     scur[0].execute("SELECT OGC_FID,length,trunk_rev_begin,trunk_rev_end,trunk_parent,trunk_child FROM pipes")
     print '################'
     for r in scur[0].fetchall():
@@ -49,14 +59,14 @@ def test():
     scur[0].execute("UPDATE pipes_view SET length = 2")
     scur[0].execute("DELETE FROM pipes_view WHERE OGC_FID = 6")
     scur[0].commit()
-    versioning.commit( wc[0], "commit 3 wc0", "dbname=epanet_test_db" )
+    spversioning.commit( [wc[0], pg_conn_info], "commit 3 wc0" )
 
     scur[0].execute("SELECT OGC_FID,length,trunk_rev_begin,trunk_rev_end,trunk_parent,trunk_child FROM pipes")
     print '################'
     for r in scur[0].fetchall():
         print r
 
-    versioning.update( wc[1], "dbname=epanet_test_db" )
+    spversioning.update( [wc[1], pg_conn_info] )
 
     scur[1].execute("SELECT OGC_FID,length,trunk_rev_begin,trunk_rev_end,trunk_parent,trunk_child FROM pipes_diff")
     print '################ diff'
