+++ test/issue485_test.py
@@ -1,23 +1,34 @@
-#!/usr/bin/python
-from .. import versioning
+#!/usr/bin/env python2
+from __future__ import absolute_import
+import sys
+sys.path.insert(0, '..')
+
+from versioningDB import versioning 
 from pyspatialite import dbapi2
+from versioningDB.sp_versioning import spVersioning
+from versioningDB.utils import Db
 import psycopg2
 import os
 import shutil
 import tempfile
 
+PGUSER = 'postgres'
+HOST = '127.0.0.1'
+
+pg_conn_info = "dbname=epanet_test_db host="+HOST+" user="+PGUSER
+
 def test():
+    spversioning = spVersioning()
+    
     test_data_dir = os.path.dirname(os.path.realpath(__file__))
     tmp_dir = tempfile.gettempdir()
 
     # create the test database
+    os.system("dropdb --if-exists -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("createdb -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -c 'CREATE EXTENSION postgis'")
 
-    os.system("dropdb epanet_test_db")
-    os.system("createdb epanet_test_db")
-    os.system("psql epanet_test_db -c 'CREATE EXTENSION postgis'")
-
-    pg_conn_info = "dbname=epanet_test_db"
-    pcur = versioning.Db(psycopg2.connect(pg_conn_info))
+    pcur = Db(psycopg2.connect(pg_conn_info))
     pcur.execute("CREATE SCHEMA epanet")
     pcur.execute("""
         CREATE TABLE epanet.junctions (
@@ -86,7 +97,7 @@ def test():
     versioning.add_branch( pg_conn_info, 'epanet', 'mybranch', 'test msg' )
 
 
-    pcur = versioning.Db(psycopg2.connect(pg_conn_info))
+    pcur = Db(psycopg2.connect(pg_conn_info))
     pcur.execute("SELECT * FROM epanet_mybranch_rev_head.junctions")
     assert( len(pcur.fetchall()) == 2 )
     pcur.execute("SELECT * FROM epanet_mybranch_rev_head.pipes")
@@ -116,14 +127,14 @@ def test():
 
     wc = tmp_dir+'/wc_multiple_geometry_test.sqlite'
     if os.path.isfile(wc): os.remove(wc)
-    versioning.checkout( pg_conn_info, ['epanet_trunk_rev_head.pipes','epanet_trunk_rev_head.junctions'], wc )
+    spversioning.checkout( pg_conn_info, ['epanet_trunk_rev_head.pipes','epanet_trunk_rev_head.junctions'], wc )
 
 
-    scur = versioning.Db( dbapi2.connect(wc) )
+    scur = Db( dbapi2.connect(wc) )
     scur.execute("UPDATE junctions_view SET GEOMETRY = GeometryFromText('POINT(3 3)',2154) WHERE OGC_FID = 1")
     scur.commit()
     scur.close()
-    versioning.commit( wc, 'moved a junction', 'dbname=epanet_test_db' )
+    spversioning.commit( [wc, pg_conn_info], 'moved a junction' )
 
     pcur.execute("SELECT ST_AsText(geometry), ST_AsText(geometry_schematic) FROM epanet_trunk_rev_head.junctions ORDER BY hid DESC")
     res = pcur.fetchall()
@@ -136,16 +147,16 @@ def test():
     # now we branch from head
     versioning.add_branch( pg_conn_info, 'epanet', 'b1', 'add branch b1' )
 
-    pcur = versioning.Db(psycopg2.connect(pg_conn_info))
+    pcur = Db(psycopg2.connect(pg_conn_info))
     pcur.execute("SELECT hid, trunk_rev_begin, trunk_rev_end, b1_rev_begin, b1_rev_end FROM epanet.junctions ORDER BY hid")
     for r in pcur.fetchall(): print r
     pcur.close()
 
     # edit a little with a new wc
     os.remove(wc)
-    versioning.checkout( pg_conn_info, ['epanet_b1_rev_head.junctions'], wc )
+    spversioning.checkout( pg_conn_info, ['epanet_b1_rev_head.junctions'], wc )
 
-    scur = versioning.Db( dbapi2.connect(wc) )
+    scur = Db( dbapi2.connect(wc) )
     scur.execute("UPDATE junctions_view SET GEOMETRY = GeometryFromText('POINT(4 4)',2154) WHERE OGC_FID = 3")
     scur.commit()
     scur.execute("PRAGMA table_info(junctions_view)")
@@ -153,9 +164,9 @@ def test():
     for r in scur.fetchall(): print r
     scur.close()
 
-    versioning.commit( wc, 'moved a junction', 'dbname=epanet_test_db' )
+    spversioning.commit( [wc, pg_conn_info], 'moved a junction')
 
-    pcur = versioning.Db(psycopg2.connect(pg_conn_info))
+    pcur = Db(psycopg2.connect(pg_conn_info))
     pcur.execute("SELECT hid, trunk_rev_begin, trunk_rev_end, b1_rev_begin, b1_rev_end FROM epanet.junctions ORDER BY hid")
     print "-----------------"
     for r in pcur.fetchall(): print r
