+++ test/multiple_geometry_test.py
@@ -1,10 +1,21 @@
-# coding = utf-8
-from .. import versioning
+#!/usr/bin/env python2
+from __future__ import absolute_import
+import sys
+sys.path.insert(0, '..')
+
+from versioningDB import versioning 
+from versioningDB.sp_versioning import spVersioning
+from versioningDB.utils import Db
 from pyspatialite import dbapi2
 import psycopg2
 import os
 import shutil
 import tempfile
+
+PGUSER = 'postgres'
+HOST = '127.0.0.1'
+
+pg_conn_info = "dbname=epanet_test_db host="+HOST+" user="+PGUSER
 
 def test():
     test_data_dir = os.path.dirname(os.path.realpath(__file__))
@@ -12,12 +23,11 @@ def test():
 
     # create the test database
 
-    os.system("dropdb epanet_test_db")
-    os.system("createdb epanet_test_db")
-    os.system("psql epanet_test_db -c 'CREATE EXTENSION postgis'")
+    os.system("dropdb --if-exists -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("createdb -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -c 'CREATE EXTENSION postgis'")
 
-    pg_conn_info = "dbname=epanet_test_db"
-    pcur = versioning.Db(psycopg2.connect(pg_conn_info))
+    pcur = Db(psycopg2.connect(pg_conn_info))
     pcur.execute("CREATE SCHEMA epanet")
     pcur.execute("""
         CREATE TABLE epanet.junctions (
@@ -86,7 +96,7 @@ def test():
     versioning.add_branch( pg_conn_info, 'epanet', 'mybranch', 'test msg' )
 
 
-    pcur = versioning.Db(psycopg2.connect(pg_conn_info))
+    pcur = Db(psycopg2.connect(pg_conn_info))
     pcur.execute("SELECT * FROM epanet_mybranch_rev_head.junctions")
     assert( len(pcur.fetchall()) == 2 )
     pcur.execute("SELECT * FROM epanet_mybranch_rev_head.pipes")
@@ -113,18 +123,19 @@ def test():
     assert( res[0][0] == 'POINT(0 0)' )
     assert( res[1][1] == 'POLYGON((0 0,2 0,2 2,0 2,0 0))' )
 
+    spversioning = spVersioning()
 
     wc = tmp_dir+'/wc_multiple_geometry_test.sqlite'
     if os.path.isfile(wc): os.remove(wc)
-    versioning.checkout( pg_conn_info, ['epanet_trunk_rev_head.pipes','epanet_trunk_rev_head.junctions'], wc )
+    spversioning.checkout( pg_conn_info, ['epanet_trunk_rev_head.pipes','epanet_trunk_rev_head.junctions'], wc )
 
 
-    scur = versioning.Db( dbapi2.connect(wc) )
+    scur = Db( dbapi2.connect(wc) )
     scur.execute("UPDATE junctions_view SET GEOMETRY = GeometryFromText('POINT(3 3)',2154)")
     scur.commit()
     scur.close()
 
-    versioning.commit( wc, 'a commit msg', 'dbname=epanet_test_db' )
+    spversioning.commit( [wc, pg_conn_info], 'a commit msg' )
 
     pcur.execute("SELECT ST_AsText(geometry), ST_AsText(geometry_schematic) FROM epanet_trunk_rev_head.junctions")
     res = pcur.fetchall()
