+++ test/partial_checkout_test.py
@@ -1,12 +1,26 @@
-#!/usr/bin/python
-from .. import versioning
+#!/usr/bin/env python2
+import sys
+sys.path.insert(0, '..')
+
+from versioningDB.pg_versioning import pgVersioning
+from versioningDB.sp_versioning import spVersioning 
+from versioningDB.versioning import historize
 
 from pyspatialite import dbapi2
 import psycopg2
 import os
 import tempfile
 
+PGUSER = 'postgres'
+HOST = '127.0.0.1'
+
+pg_conn_info = "dbname=epanet_test_db host="+HOST+" user="+PGUSER
+
 def test():
+    
+    pgversioning = pgVersioning()
+    spversioning = spVersioning()
+    
     tmp_dir = tempfile.gettempdir()
     test_data_dir = os.path.dirname(os.path.realpath(__file__))
 
@@ -15,13 +29,13 @@ def test():
         os.remove(sqlite_test_filename)
 
     # create the test database
-    os.system("dropdb epanet_test_db")
-    os.system("createdb epanet_test_db")
-    os.system("psql epanet_test_db -c 'CREATE EXTENSION postgis'")
-    os.system("psql epanet_test_db -f "+test_data_dir+"/epanet_test_db_unversioned.sql")
+    os.system("dropdb --if-exists -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("createdb -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -c 'CREATE EXTENSION postgis'")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -f "+test_data_dir+"/epanet_test_db.sql")
     
 
-    pcon = psycopg2.connect("dbname=epanet_test_db")
+    pcon = psycopg2.connect(pg_conn_info)
     pcur = pcon.cursor()
     for i in range(10):
         pcur.execute("""
@@ -38,10 +52,10 @@ def test():
     pcon.commit()
     pcon.close()
 
-    versioning.historize('dbname=epanet_test_db', 'epanet')
+    historize(pg_conn_info, 'epanet')
 
     # spatialite working copy
-    versioning.checkout("dbname=epanet_test_db",["epanet_trunk_rev_head.junctions","epanet_trunk_rev_head.pipes"], sqlite_test_filename, [[1, 2, 3], []])
+    spversioning.checkout(pg_conn_info,["epanet_trunk_rev_head.junctions","epanet_trunk_rev_head.pipes"], sqlite_test_filename, [[1, 2, 3], []])
     assert( os.path.isfile(sqlite_test_filename) and "sqlite file must exist at this point" )
 
     scon = dbapi2.connect(sqlite_test_filename)
@@ -50,9 +64,9 @@ def test():
     assert len(scur.fetchall()) ==  3
 
     # postgres working copy
-    versioning.pg_checkout("dbname=epanet_test_db",["epanet_trunk_rev_head.junctions","epanet_trunk_rev_head.pipes"], 'my_working_copy', [[1, 2, 3], []])
+    pgversioning.checkout(pg_conn_info,["epanet_trunk_rev_head.junctions","epanet_trunk_rev_head.pipes"], 'my_working_copy', [[1, 2, 3], []])
 
-    pcon = psycopg2.connect("dbname=epanet_test_db")
+    pcon = psycopg2.connect(pg_conn_info)
     pcur = pcon.cursor()
     pcur.execute("SELECT * from my_working_copy.junctions_view")
     assert len(pcur.fetchall()) ==  3
