+++ test/posgres_working_copy_bug_in_conflict_test.py
@@ -1,8 +1,18 @@
-#!/usr/bin/python
-from .. import versioning
+#!/usr/bin/env python2
+from __future__ import absolute_import
+import sys
+sys.path.insert(0, '..')
+
+from versioningDB.pg_versioning import pgVersioning
+from versioningDB.utils import Db
 import psycopg2
 import os
 import shutil
+
+PGUSER = 'postgres'
+HOST = '127.0.0.1'
+
+pg_conn_info = "dbname=epanet_test_db host="+HOST+" user="+PGUSER
 
 def prtTab( cur, tab ):
     print "--- ",tab," ---"
@@ -22,32 +32,33 @@ def test():
 
     # create the test database
 
+    versioning = pgVersioning()
     for resolution in ['theirs','mine']:
-        os.system("dropdb --if-exists epanet_test_db")
-        os.system("createdb epanet_test_db")
-        os.system("psql epanet_test_db -c 'CREATE EXTENSION postgis'")
-        os.system("psql epanet_test_db -f "+test_data_dir+"/epanet_test_db.sql")
+        os.system("dropdb --if-exists -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+        os.system("createdb -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+        os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -c 'CREATE EXTENSION postgis'")
+        os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -f "+test_data_dir+"/epanet_test_db.sql")
 
-        pcur = versioning.Db(psycopg2.connect("dbname=epanet_test_db"))
+        pcur = Db(psycopg2.connect(pg_conn_info))
 
         tables = ['epanet_trunk_rev_head.junctions', 'epanet_trunk_rev_head.pipes']
-        versioning.pg_checkout("dbname=epanet_test_db",tables, "wc1")
-        versioning.pg_checkout("dbname=epanet_test_db",tables, "wc2")
+        versioning.checkout(pg_conn_info,tables, "wc1")
+        versioning.checkout(pg_conn_info,tables, "wc2")
         print "checkout done"
 
         pcur.execute("UPDATE wc1.pipes_view SET length = 4 WHERE pid = 1")
         prtTab( pcur, "wc1.pipes_diff")
         pcur.commit()
         #pcur.close()
-        versioning.pg_commit("dbname=epanet_test_db","wc1","msg1")
+        versioning.commit([pg_conn_info,"wc1"],"msg1")
 
-        #pcur = versioning.Db(psycopg2.connect("dbname=epanet_test_db"))
+        #pcur = Db(psycopg2.connect(pg_conn_info))
 
         print "commited"
         pcur.execute("UPDATE wc2.pipes_view SET length = 5 WHERE pid = 1")
         prtTab( pcur, "wc2.pipes_diff")
         pcur.commit()
-        versioning.pg_update("dbname=epanet_test_db","wc2")
+        versioning.update([pg_conn_info,"wc2"])
         print "updated"
         prtTab( pcur, "wc2.pipes_diff")
         prtTab( pcur, "wc2.pipes_conflicts")
