+++ test/posgres_working_copy_bug_in_view_test.py
@@ -1,8 +1,18 @@
-#!/usr/bin/python
-from .. import versioning
+#!/usr/bin/env python2
+from __future__ import absolute_import
+import sys
+sys.path.insert(0, '..')
+
+from versioningDB.pg_versioning import pgVersioning
+from versioningDB.utils import Db
 import psycopg2
 import os
 import shutil
+
+PGUSER = 'postgres'
+HOST = '127.0.0.1'
+
+pg_conn_info = "dbname=epanet_test_db host="+HOST+" user="+PGUSER
 
 def prtTab( cur, tab ):
     print "--- ",tab," ---"
@@ -20,17 +30,18 @@ def prtHid( cur, tab ):
 def test():
     test_data_dir = os.path.dirname(os.path.realpath(__file__))
 
+    versioning = pgVersioning()
     # create the test database
 
-    os.system("dropdb epanet_test_db")
-    os.system("createdb epanet_test_db")
-    os.system("psql epanet_test_db -c 'CREATE EXTENSION postgis'")
-    os.system("psql epanet_test_db -f "+test_data_dir+"/epanet_test_db.sql")
+    os.system("dropdb --if-exists -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("createdb -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -c 'CREATE EXTENSION postgis'")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -f "+test_data_dir+"/epanet_test_db.sql")
 
     # chechout
-    versioning.pg_checkout("dbname=epanet_test_db",['epanet_trunk_rev_head.junctions','epanet_trunk_rev_head.pipes'], "epanet_working_copy")
+    versioning.checkout(pg_conn_info,['epanet_trunk_rev_head.junctions','epanet_trunk_rev_head.pipes'], "epanet_working_copy")
 
-    pcur = versioning.Db(psycopg2.connect("dbname=epanet_test_db"))
+    pcur = Db(psycopg2.connect(pg_conn_info))
 
     pcur.execute("UPDATE epanet_working_copy.pipes_view SET length = 4 WHERE pid = 1")
     prtTab(pcur, 'epanet_working_copy.pipes_diff')
