+++ test/versioning_base_test.py
@@ -1,10 +1,18 @@
-#!/usr/bin/python
-from .. import versioning
+#!/usr/bin/env python2
+from __future__ import absolute_import
+import sys
+sys.path.insert(0, '..')
 
+from versioningDB.sp_versioning import spVersioning
+from versioningDB.versioning import diff_rev_view_str
 from pyspatialite import dbapi2
 import psycopg2
 import os
 import tempfile
+
+PGUSER = 'postgres'
+HOST = '127.0.0.1'
+pg_conn_info = "dbname=epanet_test_db host="+HOST+" user="+PGUSER
 
 def test():
     tmp_dir = tempfile.gettempdir()
@@ -23,25 +31,26 @@ def test():
 
     # create the test database
 
-    os.system("dropdb epanet_test_db")
-    os.system("createdb epanet_test_db")
-    os.system("psql epanet_test_db -c 'CREATE EXTENSION postgis'")
-    os.system("psql epanet_test_db -f "+test_data_dir+"/epanet_test_db.sql")
+    os.system("dropdb --if-exists -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("createdb -h " + HOST + " -U "+PGUSER+" epanet_test_db")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -c 'CREATE EXTENSION postgis'")
+    os.system("psql -h " + HOST + " -U "+PGUSER+" epanet_test_db -f "+test_data_dir+"/epanet_test_db.sql")
 
+    versioning = spVersioning()
     # chechout two tables
 
     try:
-        versioning.checkout("dbname=epanet_test_db",["epanet_trunk_rev_head.junctions","epanet.pipes"], sqlite_test_filename1)
+        versioning.checkout(pg_conn_info,["epanet_trunk_rev_head.junctions","epanet.pipes"], sqlite_test_filename1)
         assert(False and "checkout from schema withouti suffix _branch_rev_head should not be successfull")
     except RuntimeError:
         pass
 
     assert( not os.path.isfile(sqlite_test_filename1) and "sqlite file must not exist at this point" )
-    versioning.checkout("dbname=epanet_test_db",["epanet_trunk_rev_head.junctions","epanet_trunk_rev_head.pipes"], sqlite_test_filename1)
+    versioning.checkout(pg_conn_info,["epanet_trunk_rev_head.junctions","epanet_trunk_rev_head.pipes"], sqlite_test_filename1)
     assert( os.path.isfile(sqlite_test_filename1) and "sqlite file must exist at this point" )
 
     try:
-        versioning.checkout("dbname=epanet_test_db",["epanet_trunk_rev_head.junctions","epanet_trunk_rev_head.pipes"], sqlite_test_filename1)
+        versioning.checkout(pg_conn_info,["epanet_trunk_rev_head.junctions","epanet_trunk_rev_head.pipes"], sqlite_test_filename1)
         assert(False and "trying to checkout on an existing file must fail")
     except RuntimeError:
         pass
@@ -55,8 +64,8 @@ def test():
     scur.execute("SELECT COUNT(*) FROM junctions")
     assert( scur.fetchone()[0] == 3 )
     scon.close()
-    versioning.commit(sqlite_test_filename1, 'first edit commit', "dbname=epanet_test_db")
-    pcon = psycopg2.connect("dbname=epanet_test_db")
+    versioning.commit([sqlite_test_filename1, pg_conn_info], 'first edit commit')
+    pcon = psycopg2.connect(pg_conn_info)
     pcur = pcon.cursor()
     pcur.execute("SELECT COUNT(*) FROM epanet.junctions")
     assert( pcur.fetchone()[0] == 3 )
@@ -65,7 +74,7 @@ def test():
 
     # add revision : edit one table and commit changes; rev = 3
 
-    versioning.checkout("dbname=epanet_test_db",["epanet_trunk_rev_head.junctions"], sqlite_test_filename2)
+    versioning.checkout(pg_conn_info,["epanet_trunk_rev_head.junctions"], sqlite_test_filename2)
 
     scon = dbapi2.connect(sqlite_test_filename2)
     scur = scon.cursor()
@@ -74,11 +83,11 @@ def test():
     #scur.execute("SELECT COUNT(*) FROM junctions")
     #assert( scur.fetchone()[0] == 3 )
     scon.close()
-    versioning.commit(sqlite_test_filename2, 'second edit commit', "dbname=epanet_test_db")
+    versioning.commit([sqlite_test_filename2, pg_conn_info], 'second edit commit')
 
     # add revision : insert one junction and commit changes; rev = 4
 
-    versioning.checkout("dbname=epanet_test_db",["epanet_trunk_rev_head.junctions"], sqlite_test_filename3)
+    versioning.checkout(pg_conn_info,["epanet_trunk_rev_head.junctions"], sqlite_test_filename3)
 
     scon = dbapi2.connect(sqlite_test_filename3)
     scur = scon.cursor()
@@ -87,11 +96,11 @@ def test():
     #scur.execute("SELECT COUNT(*) FROM junctions")
     #assert( scur.fetchone()[0] == 3 )
     scon.close()
-    versioning.commit(sqlite_test_filename3, 'insert commit', "dbname=epanet_test_db")
+    versioning.commit([sqlite_test_filename3, pg_conn_info], 'insert commit')
 
     # add revision : delete one junction and commit changes; rev = 5
 
-    versioning.checkout("dbname=epanet_test_db",["epanet_trunk_rev_head.junctions"], sqlite_test_filename4)
+    versioning.checkout(pg_conn_info,["epanet_trunk_rev_head.junctions"], sqlite_test_filename4)
 
     scon = dbapi2.connect(sqlite_test_filename4)
     scur = scon.cursor()
@@ -100,10 +109,9 @@ def test():
     #scur.execute("SELECT COUNT(*) FROM junctions")
     #assert( scur.fetchone()[0] == 3 )
     scon.close()
-    versioning.commit(sqlite_test_filename4, 'delete id=0 commit', "dbname=epanet_test_db")
+    versioning.commit([sqlite_test_filename4, pg_conn_info], 'delete id=0 commit')
 
-    pg_conn_info = "dbname=epanet_test_db"
-    select_str = versioning.diff_rev_view_str(pg_conn_info, 'epanet', 'junctions','trunk', 1,2)
+    select_str = diff_rev_view_str(pg_conn_info, 'epanet', 'junctions','trunk', 1,2)
     pcur.execute(select_str)
     res = pcur.fetchall()
     assert(res[0][0] == 'u')
@@ -111,7 +119,7 @@ def test():
     #fetchall 1 vs 2 = [
     #('u', 3, '1', 8.0, None, None, '01010000206A0800000000000000000000000000000000F03F', 2, 2, 2, 4)]
 
-    select_str = versioning.diff_rev_view_str(pg_conn_info, 'epanet', 'junctions','trunk', 1,3)
+    select_str = diff_rev_view_str(pg_conn_info, 'epanet', 'junctions','trunk', 1,3)
     pcur.execute(select_str)
     res = pcur.fetchall()
     assert(res[0][0] == 'u')
@@ -121,7 +129,7 @@ def test():
     #('u', 4, '1', 22.0, None, None, '01010000206A0800000000000000000000000000000000F03F', 3, None, 3, None),
     #('i', 3, '1', 8.0, None, None, '01010000206A0800000000000000000000000000000000F03F', 2, 2, 2, 4)]
 
-    select_str = versioning.diff_rev_view_str(pg_conn_info, 'epanet', 'junctions','trunk', 1,4)
+    select_str = diff_rev_view_str(pg_conn_info, 'epanet', 'junctions','trunk', 1,4)
     pcur.execute(select_str)
     res = pcur.fetchall()
     assert(res[0][0] == 'u')
@@ -135,7 +143,7 @@ def test():
     #('a', 5, '10', 100.0, None, None, '01010000206A08000000000000000000400000000000000000', 4, None, None, None),
     #('i', 1, '0', 0.0, None, None, '01010000206A080000000000000000F03F0000000000000000', 1, 4, None, None)]
 
-    select_str = versioning.diff_rev_view_str(pg_conn_info, 'epanet', 'junctions','trunk', 1,5)
+    select_str = diff_rev_view_str(pg_conn_info, 'epanet', 'junctions','trunk', 1,5)
     pcur.execute(select_str)
     res = pcur.fetchall()
     assert(res[0][0] == 'u')
@@ -151,7 +159,7 @@ def test():
 
     # add revision : edit one table then delete and commit changes; rev = 6
 
-    versioning.checkout("dbname=epanet_test_db",["epanet_trunk_rev_head.junctions"], sqlite_test_filename5)
+    versioning.checkout(pg_conn_info,["epanet_trunk_rev_head.junctions"], sqlite_test_filename5)
 
     scon = dbapi2.connect(sqlite_test_filename5)
     scur = scon.cursor()
@@ -161,7 +169,7 @@ def test():
     #scur.execute("SELECT COUNT(*) FROM junctions")
     #assert( scur.fetchone()[0] == 3 )
     scon.close()
-    versioning.commit(sqlite_test_filename5, 'update and delete commit', "dbname=epanet_test_db")
+    versioning.commit([sqlite_test_filename5, pg_conn_info], 'update and delete commit')
 
 
 if __name__ == "__main__":
